Stole the weareenvoy.com navigation bar
The Navbar is not a 1:1 match. The navigation bar is working, there are some small problems with mobile view. I'm working on to fix that. Class names will most likely be renamed, because they are created to match the original weareenvoy site. 

The javascript files is messy, it's mainly barba.js stuff, editing js could lead to breaking of a lot of things, don't do that.